Attempt to refactor profile page. Buggy.
