+++ client/scripts/controllers/app/login.ts
@@ -19,7 +19,6 @@ import {
 } from 'models';
 import moment from 'moment';
 import { notifyError } from 'controllers/app/notifications';
-import { networkToBase } from 'models/types';
 
 const MAGIC_PUBLISHABLE_KEY = 'pk_live_B0604AA1B8EEFDB4';
